one-stop-advisor/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
SQLAlchemy==2.0.35
python-dotenv==1.0.1
WTForms==3.1.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.4.0
pytesseract==0.3.10
gunicorn==22.0.0